# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Transport Monitoring API
  description: API REST para consultar estado de vehículos, agregados de retrasos y cuellos de botella (datos en MongoDB). Documentación para publicación con Swagger UI.
  version: 1.0.0
  contact:
    name: Proyecto Big Data

servers:
  - url: http://localhost:5000
    description: Entorno local
  - url: http://nodo1:5000
    description: Servidor del cluster

tags:
  - name: vehicles
    description: Estado de vehículos (MongoDB vehicle_status)
  - name: delays
    description: Agregados de retrasos por ruta y ventana (MongoDB route_delay_aggregates)
  - name: bottlenecks
    description: Nodos cuello de botella (MongoDB bottlenecks)
  - name: health
    description: Salud del servicio
  - name: routes
    description: Recomendación de rutas (sin IA o con modelo de predicción)

paths:
  /health:
    get:
      tags: [health]
      summary: Estado del servicio
      responses:
        '200':
          description: Servicio operativo
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  mongo: { type: string, example: connected }

  /vehicles:
    get:
      tags: [vehicles]
      summary: Listar último estado de vehículos
      parameters:
        - name: route_id
          in: query
          schema: { type: string, example: R001 }
          description: Filtrar por ruta
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Lista de estados de vehículos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/VehicleStatus' }

  /vehicles/{vehicle_id}:
    get:
      tags: [vehicles]
      summary: Último estado de un vehículo
      parameters:
        - name: vehicle_id
          in: path
          required: true
          schema: { type: string, example: V001 }
      responses:
        '200':
          description: Estado del vehículo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VehicleStatus' }
        '404':
          description: Vehículo no encontrado

  /delays:
    get:
      tags: [delays]
      summary: Agregados de retrasos por ventana
      parameters:
        - name: route_id
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Lista de agregados
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DelayAggregate' }

  /bottlenecks:
    get:
      tags: [bottlenecks]
      summary: Nodos cuello de botella (por grado)
      responses:
        '200':
          description: Lista de bottlenecks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Bottleneck' }

  /routes/recommend:
    get:
      tags: [routes]
      summary: Recomendar ruta entre dos nodos
      parameters:
        - name: from_node
          in: query
          required: true
          schema: { type: string, enum: [A, B, C, D, E] }
          description: Nodo origen
        - name: to_node
          in: query
          required: true
          schema: { type: string, enum: [A, B, C, D, E] }
          description: Nodo destino
        - name: use_ml
          in: query
          schema: { type: boolean, default: false }
          description: Si true, usa modelo de predicción de retrasos (Paso 3)
        - name: at
          in: query
          schema: { type: string, format: date-time }
          description: Fecha/hora para predicción (solo si use_ml=true)
      responses:
        '200':
          description: Ruta recomendada (path, edges, total_weight, total_time_minutes, model_used)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteRecommendation' }
        '404':
          description: No existe camino entre los nodos

components:
  schemas:
    VehicleStatus:
      type: object
      properties:
        vehicle_id: { type: string }
        route_id: { type: string }
        timestamp: { type: string, format: date-time }
        latitude: { type: number }
        longitude: { type: number }
        speed: { type: number }
        status: { type: string }
      description: Último estado conocido del vehículo (MongoDB vehicle_status)

    DelayAggregate:
      type: object
      properties:
        window_start: { type: string, format: date-time }
        window_end: { type: string, format: date-time }
        route_id: { type: string }
        total_vehicles: { type: integer }
        avg_speed: { type: number }
        delayed_count: { type: integer }
        delay_percentage: { type: number }
        max_speed: { type: number }
        min_speed: { type: number }
      description: Agregado de retrasos por ventana (MongoDB route_delay_aggregates)

    Bottleneck:
      type: object
      properties:
        node_id: { type: string }
        node_name: { type: string }
        degree: { type: integer }
        detected_at: { type: string, format: date-time }
      description: Nodo con alto grado en la red (MongoDB bottlenecks)

    RouteRecommendation:
      type: object
      properties:
        path: { type: array, items: { type: string }, description: Secuencia de nodos }
        edges: { type: array, items: { type: array, items: { type: string } }, description: Aristas (src, dst) }
        total_weight: { type: number, description: Tiempo efectivo total (ponderado por retraso) }
        total_time_minutes: { type: number, description: Tiempo base en minutos }
        model_used: { type: boolean, description: Si se usó el modelo de predicción (Paso 3) }
      description: Ruta recomendada entre origen y destino